functions/trigger_function_2/requirements.txt
functions-framework
google-cloud-functions
google-cloud-secret-manager